requirements.txt
fastapi==0.60.1
google-cloud-datastore==1.13.2
